Time: 5 ms (72.4%), Space: 44.2 MB (48.09%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -3,7 +3,7 @@ class Solution {
     if (x < 0)
       return false;
 
-    long reversed = 0;
+    int reversed = 0;
     int y = x;
 
     while (y > 0) {
